split music parsing from generic parser
